pyproject.toml
[project]
name = "drive-mcp"
version = "0.1.0"
description = "MCP server for Google Drive and OneDrive"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Henry Chien" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "msal>=1.20.0",
    "requests>=2.28.0",
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
]

[project.scripts]
drive-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
